chore: move ItemContainer to menu-item.tsx
